Establish baseline for chrome 25.0.1364.169.
Extract the release source from chrome repo using:
    gclient config https://src.chromium.org/chrome/releases/25.0.1364.169
    gclient sync --nohooks
    . build/android/envsetup.sh
    gclient runhooks

Then this is overlayed on the mips clank repo, tarball-469 branch (chrome
18.0.1025.469 tarball release).

This results in a single very large commit.

This is take-2, since my initial attempt left many old files in place.

+++ webkit/quota/quota_task.cc
@@ -1,4 +1,4 @@
-// Copyright (c) 2011 The Chromium Authors. All rights reserved.
+// Copyright (c) 2012 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -8,10 +8,12 @@
 #include <functional>
 
 #include "base/bind.h"
+#include "base/location.h"
 #include "base/message_loop.h"
 #include "base/message_loop_proxy.h"
+#include "base/single_thread_task_runner.h"
 
-using base::MessageLoopProxy;
+using base::TaskRunner;
 
 namespace quota {
 
@@ -28,11 +30,12 @@ void QuotaTask::Start() {
 
 QuotaTask::QuotaTask(QuotaTaskObserver* observer)
     : observer_(observer),
-      original_message_loop_(MessageLoopProxy::current()) {
+      original_task_runner_(base::MessageLoopProxy::current()),
+      delete_scheduled_(false) {
 }
 
 void QuotaTask::CallCompleted() {
-  DCHECK(original_message_loop_->BelongsToCurrentThread());
+  DCHECK(original_task_runner_->BelongsToCurrentThread());
   if (observer_) {
     observer_->UnregisterTask(this);
     Completed();
@@ -40,47 +43,17 @@ void QuotaTask::CallCompleted() {
 }
 
 void QuotaTask::Abort() {
-  DCHECK(original_message_loop_->BelongsToCurrentThread());
+  DCHECK(original_task_runner_->BelongsToCurrentThread());
   observer_ = NULL;
   Aborted();
 }
 
 void QuotaTask::DeleteSoon() {
+  DCHECK(original_task_runner_->BelongsToCurrentThread());
+  if (delete_scheduled_)
+    return;
+  delete_scheduled_ = true;
   MessageLoop::current()->DeleteSoon(FROM_HERE, this);
-}
-
-// QuotaThreadTask ---------------------------------------------------------
-
-QuotaThreadTask::QuotaThreadTask(
-    QuotaTaskObserver* observer,
-    scoped_refptr<MessageLoopProxy> target_message_loop)
-    : QuotaTask(observer),
-      target_message_loop_(target_message_loop) {
-}
-
-QuotaThreadTask::~QuotaThreadTask() {
-}
-
-void QuotaThreadTask::Run() {
-  target_message_loop_->PostTask(
-      FROM_HERE,
-      base::Bind(&QuotaThreadTask::CallRunOnTargetThread, this));
-}
-
-void QuotaThreadTask::CallRunOnTargetThread() {
-  DCHECK(target_message_loop_->BelongsToCurrentThread());
-  if (RunOnTargetThreadAsync())
-    original_message_loop()->PostTask(
-        FROM_HERE,
-        base::Bind(&QuotaThreadTask::CallCompleted, this));
-}
-
-bool QuotaThreadTask::RunOnTargetThreadAsync() {
-  RunOnTargetThread();
-  return true;
-}
-
-void QuotaThreadTask::RunOnTargetThread() {
 }
 
 // QuotaTaskObserver -------------------------------------------------------
